+++ templates/admin/settings.php
@@ -15,7 +15,7 @@ $yoko_lc_post_types = get_post_types( array( 'public' => true ), 'objects' );
 ?>
 
 <div class="wrap ylc-settings">
-	<h1><?php esc_html_e( 'Link Checker', 'yoko-link-checker' ); ?></h1>
+	<h1><?php esc_html_e( 'Yoko Link Checker', 'yoko-link-checker' ); ?></h1>
 
 	<?php require YOKO_LC_PLUGIN_DIR . 'templates/admin/tab-nav.php'; ?>